+++ cprof/include/cprof/activity/transfer.hpp
@@ -24,8 +24,8 @@ public:
   Transfer();
   explicit Transfer(const CUpti_ActivityMemcpy *record);
 
-  double start_ns() const;
-  double dur_ns() const;
+  size_t start_ns() const;
+  size_t dur_ns() const;
 
   std::string to_json_string() const;
   cprof::chrome_tracing::CompleteEvent chrome_complete_event() const;
